refactor: simplify logic and change function order
